+++ apps/exemple/domain/account/projectors/account.projector.ts
@@ -1,7 +1,10 @@
-import { project } from '../../app';
-import { ACCOUNT_CREATED_EVENT_TYPE } from '../../event-types';
+import { project, buildProjector } from '../../app';
+import {
+  ACCOUNT_BALANCE_CHANGED,
+  ACCOUNT_CREATED_EVENT_TYPE,
+} from '../../event-types';
 
-export const AccountProjector = [
+export const AccountProjector = buildProjector([
   project(ACCOUNT_CREATED_EVENT_TYPE, async (tx, { data: { aggregateId } }) => {
     await tx.account.create({
       data: {
@@ -11,4 +14,18 @@ export const AccountProjector = [
       },
     });
   }),
-];
+  project(
+    ACCOUNT_BALANCE_CHANGED,
+    async (tx, { data: { aggregateId, newBalance }, aggregateVersion }) => {
+      await tx.account.update({
+        where: {
+          id: aggregateId,
+        },
+        data: {
+          balance: newBalance,
+          aggregateVersion,
+        },
+      });
+    }
+  ),
+]);