+++ apps/exemple/domain/event-registry.ts
@@ -1,6 +1,10 @@
-import { AccountCreatedEvent } from './account';
-import { ACCOUNT_CREATED_EVENT_TYPE } from './event-types';
+import { AccountBalanceChangedEvent, AccountCreatedEvent } from './account';
+import {
+  ACCOUNT_BALANCE_CHANGED,
+  ACCOUNT_CREATED_EVENT_TYPE,
+} from './event-types';
 
 export type EventRegistry = {
   [ACCOUNT_CREATED_EVENT_TYPE]: AccountCreatedEvent;
+  [ACCOUNT_BALANCE_CHANGED]: AccountBalanceChangedEvent;
 };