+++ apps/exemple/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @event-craft/core@0.2.2
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @event-craft/core@0.2.1
+
 ## 0.1.1
 
 ### Patch Changes