+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @event-craft/core
 
+## 0.2.2
+
+### Patch Changes
+
+- buildEventEmitter function returns eventEmitter instance
+
+## 0.2.1
+
+### Patch Changes
+
+- Fix typing for projectors
+
 ## 0.2.0
 
 ### Minor Changes